requirements.txt
ultralytics==8.0.221
opencv-python==4.8.1.78
numpy==1.26.2
PyYAML==6.0.1
torch==2.1.1
torchvision==0.16.1
pytest==8.3.2
mlflow==2.15.1
-e .